feat: implement goals list screen with atomic design pattern
