+++ src/types/goals.ts
@@ -1,0 +1,8 @@
+export interface Goal {
+  id: string;
+  title: string;
+  description: string;
+  targetAmount: number;
+  currentAmount: number;
+  targetDate: string;
+}
